move the service call to the correct entity
