+++ custom_components/alfen_wallbox/alfen.py
@@ -210,7 +210,8 @@ class AlfenDevice:
                 if prop[ID] == api_param:
                     _LOGGER.debug(f"Set {api_param} value {value}")
                     prop[VALUE] = value
-                    return True
+                    break
+
         return False
 
     async def get_value(self, api_param):
@@ -222,7 +223,7 @@ class AlfenDevice:
         _LOGGER.debug(f"Set current limit {limit}A")
         if limit > 32 | limit < 1:
             return None
-        self.set_value("2129_0", limit)
+        await self.set_value("2129_0", limit)
 
     async def set_rfid_auth_mode(self, enabled):
         _LOGGER.debug(f"Set RFID Auth Mode {enabled}")
@@ -231,13 +232,13 @@ class AlfenDevice:
         if enabled:
             value = 2
 
-        self.set_value("2126_0", value)
+        await self.set_value("2126_0", value)
 
     async def set_current_phase(self, phase):
         _LOGGER.debug(f"Set current phase {phase}")
         if phase not in ('L1', 'L2', 'L3'):
             return None
-        self.set_value("2069_0", phase)
+        await self.set_value("2069_0", phase)
 
     async def set_phase_switching(self, enabled):
         _LOGGER.debug(f"Set Phase Switching {enabled}")
@@ -246,19 +247,19 @@ class AlfenDevice:
         if enabled:
             value = 1
 
-        self.set_value("2185_0", value)
+        await self.set_value("2185_0", value)
 
     async def set_green_share(self, value):
         _LOGGER.debug(f"Set green share value {value}%")
         if value < 0 | value > 100:
             return None
-        self.set_value("3280_2", value)
+        await self.set_value("3280_2", value)
 
     async def set_comfort_power(self, value):
         _LOGGER.debug(f"Set Comfort Level {value}W")
         if value < 1400 | value > 5000:
             return None
-        self.set_value("3280_3", value)
+        await self.set_value("3280_3", value)
 
     def __get_url(self, action) -> str:
         return "https://{}/api/{}".format(self.host, action)
