+++ custom_components/alfen_wallbox/services.yaml
@@ -10,61 +10,65 @@ set_current_limit:
   fields:
     entity_id:
       description: Name(s) of entities to change.
-      example: "alfen_wallbox.garage"
+      example: "number.wallbox_current_limit"
     limit:
       description: New limit.
+      example: 16
 
 enable_rfid_authorization_mode:
   description: Enables RFID auth mode
   fields:
     entity_id:
       description: Name(s) of entities to change.
-      example: "alfen_wallbox.garage"
+      example: "select.wallbox_authorization_mode"
 
 disable_rfid_authorization_mode:
   description: Disables RFID auth mode
   fields:
     entity_id:
       description: Name(s) of entities to change.
-      example: "alfen_wallbox.garage"
+      example: "select.wallbox_authorization_mode"
 
 set_current_phase:
-  description: Set current phase mapping to L1
+  description: Set current phase mapping
   fields:
     entity_id:
       description: Name(s) of entities to change.
-      example: "alfen_wallbox.garage"
+      example: "select.wallbox_active_load_balancing_phase_connection"
     phase:
       description: phase.
+      example: "L1"
 
 enable_phase_switching:
   description: Enable phase switching
   fields:
     entity_id:
       description: Name(s) of entities to change.
-      example: "alfen_wallbox.garage"
+      example: "switch.wallbox_enable_phase_switching"
 
 disable_phase_switching:
   description: Disable phase switching
   fields:
     entity_id:
       description: Name(s) of entities to change.
-      example: "alfen_wallbox.garage"
+      example: "switch.wallbox_enable_phase_switching"
 
 set_green_share:
   description: Set Green Share Percentage
   fields:
     entity_id:
       description: Name(s) of entities to change.
-      example: "alfen_wallbox.garage"
+      example: "number.wallbox_solar_green_share"
     value:
       description: New value.
+      example: 80
 
 set_comfort_power:
   description: Set Solar Comfort Power
   fields:
     entity_id:
       description: Name(s) of entities to change.
-      example: "alfen_wallbox.garage"
+      example: "number.wallbox_solar_comfort_level"
     value:
       description: New value.
+      example: 1400
